// data/cityCoverage.generated.ts
export const cityCoverageMeta: CityCoverageMeta = {
  "generatedAt": "2025-11-25T02:28:11.870Z",
  "overlayPath": "data/overlays/cities.generated.ts",
  "sourcePath": "data/raw/places/2025_Gaz_place_national.txt",
  "minPopulation": 0,
  "funcstatFilter": "A",
  "overlayFeatureCount": 19262,
  "sourceFeatureCount": 19469,
  "coverageRatio": 0.9884431660588628,
  "topMissing": [
    {
      "id": "0153400",
      "name": "Nauvoo town",
      "stateAbbr": "AL",
      "stateName": "Alabama",
      "population": null
    },
    {
      "id": "0159640",
      "name": "Piedmont city",
      "stateAbbr": "AL",
      "stateName": "Alabama",
      "population": null
    },
    {
      "id": "0179008",
      "name": "Vincent town",
      "stateAbbr": "AL",
      "stateName": "Alabama",
      "population": null
    },
    {
      "id": "0180256",
      "name": "Waverly town",
      "stateAbbr": "AL",
      "stateName": "Alabama",
      "population": null
    },
    {
      "id": "0432020",
      "name": "Hayden town",
      "stateAbbr": "AZ",
      "stateName": "Arizona",
      "population": null
    },
    {
      "id": "0444270",
      "name": "Marana town",
      "stateAbbr": "AZ",
      "stateName": "Arizona",
      "population": null
    },
    {
      "id": "0454050",
      "name": "Peoria city",
      "stateAbbr": "AZ",
      "stateName": "Arizona",
      "population": null
    },
    {
      "id": "0505980",
      "name": "Big Flat town",
      "stateAbbr": "AR",
      "stateName": "Arkansas",
      "population": null
    },
    {
      "id": "0518880",
      "name": "Diamondhead city",
      "stateAbbr": "AR",
      "stateName": "Arkansas",
      "population": null
    },
    {
      "id": "0535040",
      "name": "Jennette town",
      "stateAbbr": "AR",
      "stateName": "Arkansas",
      "population": null
    },
    {
      "id": "0649582",
      "name": "Mountain House city",
      "stateAbbr": "CA",
      "stateName": "California",
      "population": null
    },
    {
      "id": "0812030",
      "name": "Carbonate town",
      "stateAbbr": "CO",
      "stateName": "Colorado",
      "population": null
    },
    {
      "id": "0840550",
      "name": "Keystone town",
      "stateAbbr": "CO",
      "stateName": "Colorado",
      "population": null
    },
    {
      "id": "0875640",
      "name": "Superior town",
      "stateAbbr": "CO",
      "stateName": "Colorado",
      "population": null
    },
    {
      "id": "0877290",
      "name": "Thornton city",
      "stateAbbr": "CO",
      "stateName": "Colorado",
      "population": null
    },
    {
      "id": "1015440",
      "name": "Clayton town",
      "stateAbbr": "DE",
      "stateName": "Delaware",
      "population": null
    },
    {
      "id": "1329528",
      "name": "Fitzgerald city",
      "stateAbbr": "GA",
      "stateName": "Georgia",
      "population": null
    },
    {
      "id": "1332860",
      "name": "Gillsville city",
      "stateAbbr": "GA",
      "stateName": "Georgia",
      "population": null
    },
    {
      "id": "1336528",
      "name": "Haralson town",
      "stateAbbr": "GA",
      "stateName": "Georgia",
      "population": null
    },
    {
      "id": "1345460",
      "name": "Lavonia city",
      "stateAbbr": "GA",
      "stateName": "Georgia",
      "population": null
    }
  ]
};